Add files via upload
Implemented the MediaWiki API for Wikipedia. Fetched URLs from queries. 
Had to hardcode certain URLs for queries that returned a lot of links or lacked descriptions with detailed information. 
Also, redirects don't work so I'll have to redirect to the right URL in the next update.
Each breed has the Wikipedia's content (albeit in a very human-unreadable format). Subbreed functionality has yet to be implemented.